requirements.txt
amqp==5.2.0 ; python_version >= "3.8" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.8" and python_full_version < "3.11.3"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
backports-zoneinfo[tzdata]==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
billiard==3.6.4.0 ; python_version >= "3.8" and python_version < "4.0"
celery==5.2.7 ; python_version >= "3.8" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.8" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.8" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
django-configurations==2.3.2 ; python_version >= "3.8" and python_version < "4.0"
django-cors-headers==3.11.0 ; python_version >= "3.8" and python_version < "4.0"
django-environ==0.8.1 ; python_version >= "3.8" and python_version < "4"
django-model-utils==4.2.0 ; python_version >= "3.8" and python_version < "4.0"
django-redis==5.2.0 ; python_version >= "3.8" and python_version < "4.0"
django==4.0.5 ; python_version >= "3.8" and python_version < "4.0"
djangorestframework-simplejwt==5.0.0 ; python_version >= "3.8" and python_version < "4.0"
djangorestframework==3.13.1 ; python_version >= "3.8" and python_version < "4.0"
kombu==5.4.0 ; python_version >= "3.8" and python_version < "4.0"
prompt-toolkit==3.0.47 ; python_version >= "3.8" and python_version < "4.0"
psycopg2==2.9.3 ; python_version >= "3.8" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.8" and python_version < "4.0"
python-json-logger==2.0.7 ; python_version >= "3.8" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.8" and python_version < "4.0"
redis==5.0.8 ; python_version >= "3.8" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.8" and python_version < "3.11"
tzdata==2024.1 ; python_version >= "3.8" and python_version < "4.0" and (sys_platform == "win32" or python_version < "3.9")
vine==5.1.0 ; python_version >= "3.8" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.8" and python_version < "4.0"
